src/requirements.txt
click==8.1.7
hydra-core==1.3.2
matplotlib==3.8.0
numpy==1.26.3
pytorch_lightning==2.1.3
rich==13.7.0
scikit_learn==1.2.2
torch==2.1.2
wandb==0.16.1
